Added support for the DOM DocumentType object when constructing documents using DOMOutputter.  Specifically:
Added a new DOMAdapter.createDocument(DocType dt) method.  This creates a
DOM document with the given DocType.

Implemented the method in AbstractDOMAdapter.

Changed DOMOutputter to call the new method for documents that have a
DOCTYPE declaration.  It works for namespaces now too.

While I was at it, fixed a bug where Attr output(Attribute) wasn't using
JAXP.

It's odd the DOCTYPE doesn't appear to get recorded at all when using
Xerces 1.2, but it works fine for Xerces 1.3 and the Crimson bundled
with JAXP 1.1.  Xerces 1.2 bug?

-jh-