Refactored the JAXP model for dealing with DOM.  We now have a new class JAXPDOMAdapter that contains all logic for interacting with JAXP.  This greatly simplifies DOMBuilder and DOMOutputter and allows third parties to use JAXP easily without stealing code from these classes.
As part of this change DOMAdapter methods now may throw Exception instead of
IOException.  The bad news: This is an incompatible change and there's no
way to avoid it.  The good news: adapter classes aren't for general use so
this shouldn't bother anyone using the standard JDOM classes.  DOMBuilder
and DOMOutputter have the same API as always.

This code works on all my JDOM -> DOM -> JDOM tests with Xerces 1.2, Xerces
1.3, and Crimson with JAXP 1.1.  Course, there may be some end conditions I
haven't caught.  This was a fair amt of refactoring.

-jh-