Added two new Document constructors:
Document(List)
Document(List, DocType)

This allows a document to be created from a List of content.  Should make
it easier for people to create a document with comments or PIs ahead of
the root element.

Also added a check here and in setMixedContent() so that only one elt is
allowed in the list as the root.

This has been a TODO item for some time.

-jh-