Fixed Element.setMixedContent() so it now checks object types and sets each added object's parent.  This was an XXX comment and was also pointed out recently by Will Conant <will_conant@hotmail.com>.
Also improved documentation on Element and Document to note that
"In event of an exception the content may be partially added."
We have two choices here: to pre-check for error conditions and thus make
the set operation appear "atomic" or to allow an error event to leave
the list partially changed.  It's faster to leave the list partially
changed and adding wrongly typed objects is probably a catastrophic
failure, so I implemented without the pre-check.

-jh-