Major internal changes to XMLOutputter, and some public and protected method signature changes.
The code is now much cleaner with regard to whitespace handling, and the
"text trim" printing behaves a whole lot better.  In fact, the following
settings now pretty print on output even if the doc contains some content
with whitespace from a file and some content built in memory with no
whitespace!

setTrimText(true);
setIndent("  ");
setNewlines(true);

It may not perfectly match everyone's idea of "pretty" but in my mind
it's "pretty" good.  :-)

BTW, the default settings (no trim, no indent, no newlines) still
behave the exact same as before and print the document raw.

Among the structural changes: The printXXX() methods now only print
XXX and don't concern themselves with whitespace around the XXX item.
For example, printComment() prints only the comment from <!-- to --> and
doesn't try to print the new lines and indention before and after.
The only place where whitespace handling occurs is printDocument()
for document-level whitespace and printElementContent() for the whitespace
between an element's tags.  But in both these cases the whitespace is
naturally within the item being printed.

The signature changes:

* Deprecated "public setIndentLevel()" method.  Having a global indent is
  better done with a stacked FilterOutputStream.  The method is now empty.
  It'll be removed after beta7.
* Deprected "public setPadText()" method.  It's not needed with the
  current output mechanism.  The method is now empty.  It'll be removed
  after beta7.
* Changed the "protected printXXX()" methods to have a new signature
  without the "int indentLevel" final parameter.  Didn't bother with
  deprecation here.

Thanks to Mark Roder <mroder@techies.com> for raising the issue that got
this started.

-jh-