Modified the addNamespaceDeclaration() method to now ensure the prefix doesn't collide with an existing prefix on the element.
Also went back and improved the addAttribute() method to avoid the
prefix check if the prefix is "", as an easy and common optimization.

Added a new constructor in IllegalAddException to account for a Namespace
illegally added:

public IllegalAddException(Element base, Namespace added, String reason)

Updated the TODO to reflect this work.

-jh-