Added parentage to DocType so it now has a public getDocument() method that points at its holding Document.  It also has a protected setDocument() method that is used by Document to manage its parentage.
Added notes to the Javadoc of methods that return List that the List
returned is in document order.  There was some doubt about that earlier,
so best to be clear.

Changed elt.setAttribute() so it now replaces any existing attribute with
the same name instead of throwing an IllegalAddException about a
collision.  This was part of the Javadoc, and the impl just had to
catch up.

Added a new constructor to IllegalAddException to handle the case when a
DocType is added with a pre-existing parent:
 public IllegalAddException(Document base, DocType added, String reason)

Updated TODO to reflect these changes.

-jh-