Added JDOMSource and JDOMResult to the core in the org.jdom.transform package.  They've been peer reviewed in jdom-contrib without incident, and consensus appears to be that it's better to have these in the core now than to wait a potentially very long time for JAXP.next.
I added Crimson, JAXP, and Xalan JARs to the lib directory to support the
transform functionality.  I also added a README in the lib directory to
explain what each JAR does.  :-)

I'm considering making Crimson the default parser to save download size,
but I'm hesistant because Xerces is more mature and because Crimson
outputs that nasty JIT warning on JDK 1.2.2 on Windows which is probably
the most popular platform.  Half the people are going to think it's a
JDOM issue and quit using JDOM.

I altered the build scripts to account for needing jaxp.jar in the
classpath during the build.

I added the XSLTransform sample to the samples directory along with the
supporting catalog.* files.  I'll remove them from jdom-contrib soon.

And finally I'm checking in the TODO.txt now that so much has been done
today.

-jh-