Integrated a patch from "Taylor, Patrick" <ptaylor@rational.com> to make the javadocs build with proper links to external classes.  Here's the email describing the patch:
The following patch to build.xml causes all javadoc references to external
classes to be generated as hyperlinks to the J2SE1.4 javadocs for those
classes.  This is done simply by using the -link option to the javadoc tool.
See the following page for details:

http://java.sun.com/j2se/1.4/docs/tooldocs/win32/javadoc.html#link

Using the 1.4 doc (rather than an earlier version of the J2SE) allows all
links to be resolved, as the DOM/SAX/etc packages are now included in
the J2SE.

-jh-