Fixed strange compiler error
Reference to assertEquals is ambiguous. It is defined in void assertEquals(java.lang.String, long, long) and void assertEquals(java.lang.String, int, int).

by upcasting a few ints to longs

(Not sure if this came up during the upgrade to JUnit 3.7 or because I'm using a 1.3 compiler with different ambiguity rules.)