Integrating a patch from Laurent Bihanic <laurent.bihanic@atosorigin.com> sent to jdom-interest on 27 Sept 2001.  Below is his text.
-jh-

Attached is a enhancement to SAXOutputter to bring is closer to XMLReader by
supporting all the methods defined by XMLReader except the 2 methods parse().
With these changes, implementing an XMLReader or XMLFilter from SAXOutputter
is  a lot easier: It is now possible to derive from SAXOutputter instead of
writing an adapter class.
JDOMSource has been updated to demonstrate how easy it is!

Changes include:
  - new public no-argument constructor
  - support for DeclHandler (setter and getter methods)
  - getter methods for ContentHandler, DTDHandler, ErrorHandler,
EntityResolver and LexicalHandler
  - set/getFeature and set/getProperty methods

Laurent