This reflects Laurent Bihanic's changes. Here is the text of the mail from him:
Events are fired from either the internal subset (expandEntities set to false
in SAXBuilder) or the DTD itself (expandEntities set to true). Events are
fired only if either a DTDHandler or a DeclHandler are registered on SAXOutputter.

A modified version of SAXHandler is also included. It is based on Phil's
version (mail "Re: External Entities", dated Nov. 6th) plus a patch to prevent
comments from the DTD to be added to the internal subset when expandEntities
is set to false. Alex Rosen's patch should probably be applied too.

This version has been tested with Xerces 1.4.3 and the latest JDOM from CVS.
Please note that it does not work with Crimson (1.1.1 and 1.1.3) nor Xerces
2.0.0 beta3 due to bugs that prevent these parsers to reparse the internal
subset they have generated.

Please note that this change in SAXOutputter does NOT fix the problem I wanted
to fix: Have XSLT processor recognize ID attributes in JDOM document.
It appear that neither Xalan nor Saxon use the DTD events to track ID
attributes: Both only rely on the attribute type provided in startElement's
Attributes