More changes from Laurent Bihac. This batch supports attribute typing, and   his message on the changes is included here:
Attached is an implementation to support attribute types in JDOM.
It has been validated against both Xalan and Saxon to check that they both now
recognized the ID type attributes when SAX events are fired through the
JDOMSource.
It includes handling of all known idiosyncrasies of Crimson 1.1.x, Xerces
1.4.x and Xerces 2.0.0-betaX regarding this aspect (see comments in code).

Attribute types are modeled as integer values in Attribute. SAXHandler and
SAXOutputter include methods to convert from/to SAX type strings. I added 2
methods to JDOMFactory to support the new Attribute constructors that take an
attribute type as argument.
As I don't know DOM, I haven't done any work in this direction!

Finally, something which has nothing to do with the above, I added support for
SAX Locator in SAXHandler with get/setDocumentLocator methods,
getDocumentLocator aims at allowing the JDOMFactories impl to get access to
parse location information. It is up to the application responsibility to pass
a reference to SAXHandler to their factory.