This is the long-awaited check-in that adds FilterList functionality to JDOM!  Now instead of using the slow and broken PartialList to make lists live, we'll be using a faster and smarter FilterList mechanism.
This check-in also integrates the Text class that has been
developed but previously not actually used.

It's a major change and has taken a long time to design, develop,
and test in the jdom-wip workspace.  Thanks to Alex Rosen, Bradley
S. Huffman, and Jools Enticknap for doing all the coding work.
The code I'm integrating here is absolutely beautiful!

Added classes:
  AttributeFilter
  DocumentContentFilter
  ElementContentFilter
  ElementFilter
  Filter
  FilterList

Removed classes:
  PartialList

Integrated classes:
  Text

Added methods:
  CDATA.getTextTrim()
  CDATA.getTextNormalize()
  CDATA.append(String)
  CDATA.append(CDATA)
  CDATA.getParent()
  CDATA.getDocument()
  CDATA.detach()
  Element.addContent(Text)
  Element.removeContent(Text)

This check-in also removes almost all the methods that were
deprecated in beta7.  If you were getting deprecation warnings
before, they're going to be errors now.  Heads up.

This isn't a simple change, so for those who don't want to take the
leap just yet, the CVS tag "jdom_prefilter" marks the codebase before
this check-in.  I hope everyone tries out the new code though.  It
should be faster and consume fewer resources to traverse a tree, and
all the live list methods are implemented now too!

Please report your experiences to the jdom-interest list.

-jh-