Integrating patch from Jeff Turner <jeff@socialchange.net.au> so that the JDOM versioning will make more sense.  Below is his email.
-jh-

Currently, any 'jdom.jar' built from CVS will claim that it is beta 7,
when one interrogates it via 'java -jar'. This effectively renders the
whole version reporting system useless, as a jdom.jar from 6 months ago
look exactly like one produced yesterday.

This may be a symptom of a deeper misunderstanding. Is "beta 7" a point
in time, or an ongoing period of time (about 6 months)? Are we "in the
7th beta", or has it been and gone?

I'll assume a beta is a "point in time", as this is what users believe,
and the website indicates ("JDOM Beta 7 was released!").

In which case, some terminology is needed for where JDOM is, now.

The attached patch adopts the <next version>-dev approach. Once applied,
'java -jar jdom.jar' will report:

JDOM version 1.0beta8-dev, built January 6 2002
Copyright 2000-2001, Jason Hunter and Brett McLaughlin
...

Does that sound reasonable?

If so, the implication is that when beta 8 arrives, the distro builder
must update the version flag to "beta 8", do the build, then immediately
change the flag to "beta 9-dev".