Incorporating a great patch from elharo.  In his words...
--
Here's the diffs for the changes. They don't seem to break anything
that wasn't broken before. Mostly this adds verification of the
system and public IDs in both DocType and EntityRef, the root element
name in DocType, and the entity name in EntityRef. Failing these
causes the throwing of various runtime exceptions
(IllegalDataException and IllegalNameexception) so this shouldn't
break anybody's code.

It also allows null to be used for either system or public IDs in
DocType to differentiate a missing ID from the empty string. This
necessitated some small changes in some input and output classes.

I added one new Constructor to EntityRef to allow it to have a system
ID without a public ID.

I also added a setElementName() method in DocType.

I added checkSystemLiteral() and checkPublicID() methods in Verifier.

I made checkXMLName() in Verifier public so EntityRef and DocType
could use it. (They do allow colons in their names which
checkElementName() forbids.) I don't think this will cause a problem,
but if it does it could be package private.

Finally I fixed a few small errors I noticed in the doc comments here
and there.
--

The only thing I didn't take was the change that allowed colons in XML
element names, since I'm thinking that probably causes more problems
than it solves.  I left elharo's comment in there, and we can always
make the one-line change later after the j-i discussion on the issue.

-jh-