Patch from elharo to handle skipped entities.  In his words...
--
The following patch will add support for skipped entities to SAXHandler
in the event that the parser is not resolving external entities. I''ve
verified
that it compiles and doesn't break any tests. However, I'm not totally
fluent on the current logic of SAXHandler and org.jdom.input, so please
double check my skippedEntity() method for any obvious mistakes.

I also made some corrections to the JavaDoc for processsingInstruction().
--

-jh-