Added the ability for DOMOutputter to set the internal DTD subset on DOM implementations which provide that ability, specifically Xerces 1.4.4+. This is accomplished via a new method in AbstractDOMAdapter:
protected void setInternalSubset(DocumentType dt, String s)

Which by default uses reflection to call setInternalSubset() on the
DocumentType.  That's not a standard DOM ability, but some parsers allow
it, so we try to leverage it using reflection.  If another parser provides
the ability a different way, the method can be overridden.

-jh-