Patch from Alex Rosen.
-jh-

On 11/5/01, we changed SAXBuilder so that saxDriverClass no longer gets set
after creating the parser using JAXP (see
http://www.servlets.com/archive/servlet/ReadMsg?msgId=160131&listName=jdom-i
nterest). So in our error messages, we can no longer rely on saxDriverClass
to tell the user which parser we're complaining about - we have to actually
get the parser's class each time.

Also - we are still setting saxDriverClass, if JAXP is not available. I
assume that this is an OK thing to do? It should only be a problem if JAXP
is not initially available to your SAXBuilder, but then it become available
the second time. Seems unlikely.

Alex