Adding a "scanner" example in the jdom-contrib module from Laurent.  From his readme:
ElementScanner is a SAX filter that uses XPath-like expressions to select
element nodes to build and notifies listeners when these elements becomes
available during the SAX parse.

ElementScanner does not aim at providing a faster parsing of XML documents.
Its primary focus is to allow the application to control the parse and to
consume the XML data while they are being parsed.  ElementScanner can be
viewed as a high-level SAX parser that fires events conveying JDOM elements
rather that XML tags and character data.
ElementScanner only notifies of the parsing of element nodes and does not
support reporting the parsing of DOCTYPE data, processing instructions or
comments except for those present within the selected elements.
Applications needing such data shall register a specific SAX ContentHandler
on ElementScanner to receive them in the form of raw SAX events.

--

I'm also checking in the Jaxen library which it depends on.  The version
is Jaxen 1.0 beta8.