Patch from Alex Rosen to fix a bug reported by Patrick McFarlane regarding moving attributes.  In Alex's words:
I fixed the problem in AttributeList - one of the remove() methods wasn't
detaching the parent. These patches also contain a couple of very minor
fixes:

- In ContentList, rearranged the instanceof tests to be more in order of
expected frequency (specifically I moved up the check for Text nodes higher
up). May result in a tiny speedup.

- Pet peeve: The argument to AttributeList.removeParent() should *always* be
an Attribute, or else something's very wrong with the JDOM code; we were
checking the object's type (in order to avoid the ClassCastException), but
doing nothing in the error case. This is worse than not checking - it was
making in fail silently. I removed the check, so now it would throw a
ClassCastException. (You could also keep the check and throw a different,
more helpful exception. I've never been a fan of that - this exception would
indicate a serious bug in the JDOM code itself, not the use of JDOM. I don't
know if it's worth the added code and string size...)

-jh-