Patch from elharo to improve how builders handle exceptions.
Warning: There's a lot of code that will need to change because of
this.  It's an easy change though.  Builder calls are now going
to need to catch IOException.

In his words:

--
I've attached patches for the four classes affected by the decision to
allow build() to throw IOException. I've made separate patch files for
each of the classes because my last mega-patch had troubles being
applied. The four files are:

SAXBuilder
SAXHandler
JDOMResult
TestVerifier

All the tests that ran before still run after these four patches are
applied. This is going to affect a lot of client code, but in my initial
tests not as much as I originally expected.

The basic idea of these patches is to rationalize the excepiton
throwing. This patch does several things:

1. Allows IOExceptions to propagate unchanged from the builder.
2. Converts all SAXExceptions to JDOMExceptions
3. Removes a number of "throws Exception" and "throws IOException"
clauses where no exception was being thrown.
4. Changes some "throws Exception" and "throws IOException" clauses to
"throws JDOMException"
5. Allows unexpected runtime exceptions (e.g. NullPointerException) to
propagate unchanged.

I think this makes sense. The basic idea is that SAXBuilder should throw
an IOException for an I/O error, a JDOMException for an XML problem, and
that unexpected runtime exceptions should not be hidden. I fully expect
that this will break existing code, but the sooner we do it the less
pain it will cause. As always more testing is appreciated.
--

I also updated the Count sample which needed to catch IOException.

-jh-