Patch from Brad to fix the Javadoc warnings under JDK 1.4.  In his words:
Javadoc patch to get rid of the nasty warnings caused by the change in
the definition of breakiterator.  Still have 3 warnings in XMLOutputter
cause by the use of a period before the true end of the first sentence.

--

I also changed the pm@jdom address in the license to request@jdom since
pm@jdom is getting nothing but virus email.  I'm going to /dev/null the pm
address.

-jh-