Patch from elharo:
The attached patch adds @throws clauses for runtime exceptions thrown by
the constructor, setText(), and appendText() methods in Text. It changes
the JavaDoc only and does not touch the actual code.

This is the last patch of this nature for org.jdom unless someone
notices an exception I missed. All the other classes (EntityRef, etc.)
seem to correctly document their runtime exceptions.

-jh-