Patch from elharo:
The attached patch fixes a problem with the clone() method that would
result in CDATA sections being cloned as Text objects rather than CDATA
objects. It is now necessary to check for CDATA before checking for Text
because CDATA is a subclass of Text.

It also corrects one minor spelling error in the comments.

-jh-