Patch from elharo, in his words:
I've attached three patches for the Namespace, Verifier, and
TestVerifier classes that have the net effect of removing the special
treatment of xml:space and xml:lang from JDOM. Henceforth, they will be
treated as will any other attributes in a namespace. This does change
JDOM's behavior, but I hope not too many people are relying on this.
(Indeed, it's hard to imagine they are since previously JDOM was
inconsistent depending on whether or not these attributes were created
in memory or parsed from a file and we didn't get any bug reports on
that.) The API has not changed at all.

The test case had to be changed because the previous tests checked to
see that these attributes were special cased. We've chosen to reverse
that decision so the test had to be reversed as well.

CVS was acting up on me, so please double check the patches. In
particular, make sure the Namespace patch doesn't wipe out any recent
changes from anyone else.

-jh-